pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "antsxmm"
dynamic = ["version"]
description = "A streamlined wrapper for ANTsPyMM tailored for antsxbids BIDS output"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "ANTsX Community"}
]
dependencies = [
    "antspymm>=1.1.0",
    "antspyt1w",
    "pandas>=1.0.0",
    "tqdm",
    "click",
    "antspyx",
    "importlib-metadata; python_version<'3.8'"
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "pytest-mock"
]

[project.scripts]
antsxmm = "antsxmm.pipeline:main"

[tool.setuptools.packages.find]
include = ["antsxmm*"]

[tool.setuptools_scm]
write_to = "antsxmm/_version.py"
fallback_version = "0.0.0-dev"

[tool.pytest.ini_options]
testpaths = ["tests"]
